--- Lesson04-Spring-Boot/src/main/java/ru/rrenat358/controllers/MainController.java
@@ -1,27 +1,0 @@
-package ru.rrenat358.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
-import ru.rrenat358.data.Student;
-import ru.rrenat358.repositories.StudentRepository;
-
-import java.util.List;
-
-@RestController
-public class MainController {
-
-    private StudentRepository studentRepository;
-
-    public MainController(StudentRepository studentRepository) {
-        this.studentRepository = studentRepository;
-    }
-
-    @GetMapping("/students")
-    public List<Student> getAllStudents() {
-        return studentRepository.getAllStudents();
-    }
-
-
-}
