+++ Lesson04-Spring-Boot/src/main/java/ru/rrenat358/controllers/StudentController.java
@@ -1,0 +1,35 @@
+package ru.rrenat358.controllers;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import ru.rrenat358.data.Student;
+import ru.rrenat358.repositories.StudentRepository;
+import ru.rrenat358.services.StudentService;
+
+import java.util.List;
+
+@RestController
+public class StudentController {
+    private StudentService studentService;
+
+    public StudentController(StudentService studentService) {
+        this.studentService = studentService;
+    }
+
+    @GetMapping("/students")
+    public List<Student> getAllStudents() {
+        return studentService.getAllStudents();
+    }
+
+    @GetMapping("/students/delete/{id}")
+    public void deleteById(@PathVariable Long id) {
+        studentService.deleteById(id);
+    }
+
+    @GetMapping("/students/change_score")
+    public void changeScore(@RequestParam Long studentId, @RequestParam Integer delta) {
+        studentService.changeScore(studentId, delta);
+    }
+}
